Refactor Laske Exporter to own file
